requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
mysql==0.0.2
mysql-connector-python==8.0.17
mysqlclient==1.4.4
Pillow==6.1.0
protobuf==3.9.1
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.8
urllib3==1.25.3
